% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationFilter.R
\docType{class}
\name{AnnotationFilter-class}
\alias{.AnnotationFilter}
\alias{AnnotationFilter-class}
\alias{SymbolFilter}
\alias{SymbolFilter-class}
\alias{condition}
\alias{condition,AnnotationFilter-method}
\alias{condition<-}
\alias{condition<-,AnnotationFilter-method}
\alias{show,AnnotationFilter-method}
\alias{supportedFilters}
\alias{value}
\alias{value,AnnotationFilter-method}
\alias{value<-}
\alias{value<-,AnnotationFilter-method}
\title{Filters for annotation objects

These functions are used to create filters for annotation
resources.

\code{supportedFilters()} lists all defined filters; filters are
only available for tables containing the field on which the filter
acts.}
\usage{
supportedFilters()

\S4method{show}{AnnotationFilter}(object)

\S4method{condition}{AnnotationFilter}(object)

\S4method{condition}{AnnotationFilter}(object) <- value

\S4method{value}{AnnotationFilter}(object)

\S4method{value}{AnnotationFilter}(object) <- value
}
\arguments{
\item{object}{An \code{AnnotationFilter} object}

\item{value}{The value for the object.}
}
\description{
\code{condition,condition<-} get or set the \code{condition}
value for the filter \code{object}.

\code{value,value<-} get or set the \code{value} for the filter
\code{object}.
}
\examples{
supportedFilters()

## Create a SymbolFilter to filter on a gene's symbol.
sf <- SymbolFilter("BCL2")

}

