% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationFilter.R
\name{GenenameFilter}
\alias{GenenameFilter}
\alias{GenenameFilter-class}
\title{DEPRECATED Gene name filter}
\usage{
GenenameFilter(value, condition = "==", not = FALSE)
}
\arguments{
\item{value}{\code{character()} value for the filter}

\item{condition}{\code{character(1)} defining the condition to be
used in the filter. One of \code{"=="}, \code{"!="}, \code{"startsWith"}, \code{"endsWith"}
or \code{"contains"}. Default condition is \code{"=="}.}

\item{not}{\code{logical(1)} whether the \code{AnnotationFilter} is negated.
\code{TRUE} indicates is negated (!). \code{FALSE} indicates not
negated. Default not is \code{FALSE}.}
}
\value{
The constructor function return a \code{GenenameFilter}.
}
\description{
The \code{GenenameFilter} class and functions are deprecated. Please use the
\code{\link[=GeneNameFilter]{GeneNameFilter()}} instead.
}
