% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationFilter.R
\docType{methods}
\name{AnnotationFilter}
\alias{AnnotationFilter}
\alias{CdsStartFilter}
\alias{CdsEndFilter}
\alias{ExonIdFilter}
\alias{ExonNameFilter}
\alias{ExonStartFilter}
\alias{ExonEndFilter}
\alias{ExonRankFilter}
\alias{GeneIdFilter}
\alias{GenenameFilter}
\alias{GeneBiotypeFilter}
\alias{GeneStartFilter}
\alias{GeneEndFilter}
\alias{EntrezFilter}
\alias{SymbolFilter}
\alias{TxIdFilter}
\alias{TxNameFilter}
\alias{TxBiotypeFilter}
\alias{TxStartFilter}
\alias{TxEndFilter}
\alias{ProteinIdFilter}
\alias{UniprotFilter}
\alias{SeqNameFilter}
\alias{SeqStrandFilter}
\alias{AnnotationFilter-class}
\alias{CharacterFilter-class}
\alias{IntegerFilter-class}
\alias{CdsStartFilter-class}
\alias{CdsEndFilter-class}
\alias{ExonIdFilter-class}
\alias{ExonNameFilter-class}
\alias{ExonStartFilter-class}
\alias{ExonEndFilter-class}
\alias{ExonRankFilter-class}
\alias{GeneIdFilter-class}
\alias{GenenameFilter-class}
\alias{GeneBiotypeFilter-class}
\alias{GeneStartFilter-class}
\alias{GeneEndFilter-class}
\alias{EntrezFilter-class}
\alias{SymbolFilter-class}
\alias{TxIdFilter-class}
\alias{TxNameFilter-class}
\alias{TxBiotypeFilter-class}
\alias{TxStartFilter-class}
\alias{TxEndFilter-class}
\alias{ProteinIdFilter-class}
\alias{UniprotFilter-class}
\alias{SeqNameFilter-class}
\alias{SeqStrandFilter-class}
\alias{supportedFilters}
\alias{show,AnnotationFilter-method}
\alias{show,CharacterFilter-method}
\alias{show,IntegerFilter-method}
\alias{show,GRangesFilter-method}
\alias{condition,AnnotationFilter-method}
\alias{condition}
\alias{value,AnnotationFilter-method}
\alias{value}
\alias{GRangesFilter-class}
\alias{.GRangesFilter}
\alias{GRangesFilter}
\alias{AnnotationFilter}
\alias{supportedFilters,missing-method}
\title{Filters for annotation objects}
\usage{
CdsStartFilter(value, condition = "==")
CdsEndFilter(value, condition = "==")
ExonIdFilter(value, condition = "==")
ExonNameFilter(value, condition = "==")
ExonRankFilter(value, condition = "==")
ExonStartFilter(value, condition = "==")
ExonEndFilter(value, condition = "==")
GeneIdFilter(value, condition = "==")
GenenameFilter(value, condition = "==")
GeneBiotypeFilter(value, condition = "==")
GeneStartFilter(value, condition = "==")
GeneEndFilter(value, condition = "==")
EntrezFilter(value, condition = "==")
SymbolFilter(value, condition = "==")
TxIdFilter(value, condition = "==")
TxNameFilter(value, condition = "==")
TxBiotypeFilter(value, condition = "==")
TxStartFilter(value, condition = "==")
TxEndFilter(value, condition = "==")
ProteinIdFilter(value, condition = "==")
UniprotFilter(value, condition = "==")
SeqNameFilter(value, condition = "==")
SeqStrandFilter(value, condition = "==")

\S4method{condition}{AnnotationFilter}(object)

\S4method{value}{AnnotationFilter}(object)

GRangesFilter(value, condition = "overlapping", feature = "gene")

\S4method{supportedFilters}{missing}(object)
}
\arguments{
\item{object}{An \code{AnnotationFilter} object.}

\item{value}{Value for the filter. For \code{GRangesFilter} a
\code{\link[GenomicRanges]{GRanges}} object.}

\item{condition}{character(1) defining the condition to be used in
the filter.  For numeric/integer filters one of \code{"=="},
\code{"!="}, \code{">"}, \code{"<"}, \code{">="} and
\code{"<="}. For character filter/values \code{"=="},
\code{"!="}, \code{"startsWith"} and \code{"endsWith"} are
allowed. Default condition is "==". For \code{GRangesFilter} it
can be \code{"within"} (for the feature to be completely within
the range) or \code{"overlapping"}, for the feature to be
(partially) overlapping with the range.}

\item{feature}{\code{character(1)} defining on what feature the
\code{GRangesFilter} should be applied. Choices could be
\code{"gene"}, \code{"tx"} or \code{"exon"}.}
}
\description{
The filters extending the base \code{AnnotationFilter} class
represent a simple filtering concept for annotation resources.
Each filter object is thought to filter on a single (database)
table column based on the provided values and the defined
condition.

Filter instances should be created using the constructor functions (e.g.
\code{GeneIdFilter}) and not by calls to \code{new}.

\code{supportedFilters()} lists all defined filters. Packages using
\code{AnnotationFilters} should implement the \code{supportedFilters} for
their annotation resource object (e.g. for \code{object = "EnsDb"} in the
\code{ensembldb} package) to list all supported filters for the specific
resource.

\code{condition()} get the \code{condition} value for
    the filter \code{object}.

\code{value()} get the \code{value} for the filter
    \code{object}.
}
\details{
By default filters are only available for tables containing the
field on which the filter acts (i.e. that contain a column with the
name matching the value of the \code{field} slot of the
object). See the vignette for a description to use filters for
databases in which the database table column name differs from the
default \code{field} of the filter.
}
\examples{
## filter by GRanges
GRangesFilter(GenomicRanges::GRanges("chr10:87869000-87876000"))
## Create a SymbolFilter to filter on a gene's symbol.
sf <- SymbolFilter("BCL2")
sf

## Create a GeneStartFilter to filter based on the genes' chromosomal start
## coordinates
gsf <- GeneStartFilter(10000, condition = ">")
gsf

supportedFilters()
}
