% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationFilter.R
\docType{methods}
\name{supportedFilters}
\alias{supportedFilters}
\alias{show,AnnotationFilter-method}
\title{Filters for annotation objects}
\usage{
supportedFilters()

\S4method{show}{AnnotationFilter}(object)
}
\arguments{
\item{object}{An \code{AnnotationFilter} object}
}
\description{
These functions are used to create filters for annotation
resources.
}
\details{
\code{supportedFilters()} lists all defined filters; filters are
only available for tables containing the field on which the filter
acts.
}
\examples{
supportedFilters()

}
