% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationFilter.R
\docType{methods}
\name{supportedFilters}
\alias{AnnotationFilter-class}
\alias{CdsEndFilter}
\alias{CdsEndFilter-class}
\alias{CdsStartFilter}
\alias{CdsStartFilter-class}
\alias{EntrezFilter}
\alias{EntrezFilter-class}
\alias{ExonEndFilter}
\alias{ExonEndFilter-class}
\alias{ExonIdFilter}
\alias{ExonIdFilter-class}
\alias{ExonNameFilter}
\alias{ExonNameFilter-class}
\alias{ExonRankFilter}
\alias{ExonRankFilter-class}
\alias{ExonStartFilter}
\alias{ExonStartFilter-class}
\alias{GeneBiotypeFilter}
\alias{GeneBiotypeFilter-class}
\alias{GeneEndFilter}
\alias{GeneEndFilter-class}
\alias{GeneIdFilter}
\alias{GeneIdFilter-class}
\alias{GeneStartFilter}
\alias{GeneStartFilter-class}
\alias{GenenameFilter}
\alias{GenenameFilter-class}
\alias{ProteinIdFilter}
\alias{ProteinIdFilter-class}
\alias{SeqNameFilter}
\alias{SeqNameFilter-class}
\alias{SeqStrandFilter}
\alias{SeqStrandFilter-class}
\alias{SymbolFilter}
\alias{SymbolFilter-class}
\alias{TxBiotypeFilter}
\alias{TxBiotypeFilter-class}
\alias{TxEndFilter}
\alias{TxEndFilter-class}
\alias{TxIdFilter}
\alias{TxIdFilter-class}
\alias{TxNameFilter}
\alias{TxNameFilter-class}
\alias{TxStartFilter}
\alias{TxStartFilter-class}
\alias{UniprotFilter}
\alias{UniprotFilter-class}
\alias{condition}
\alias{condition,AnnotationFilter-method}
\alias{show,AnnotationFilter-method}
\alias{supportedFilters}
\alias{value}
\alias{value,AnnotationFilter-method}
\title{Filters for annotation objects}
\usage{
CdsStartFilter(value, condition = "==")
CdsEndFilter(value, condition = "==")
ExonIdFilter(value, condition = "==")
ExonNameFilter(value, condition = "==")
ExonRankFilter(value, condition = "==")
ExonStartFilter(value, condition = "==")
ExonEndFilter(value, condition = "==")
GeneIdFilter(value, condition = "==")
GenenameFilter(value, condition = "==")
GeneBiotypeFilter(value, condition = "==")
GeneStartFilter(value, condition = "==")
GeneEndFilter(value, condition = "==")
EntrezFilter(value, condition = "==")
SymbolFilter(value, condition = "==")
TxIdFilter(value, condition = "==")
TxNameFilter(value, condition = "==")
TxBiotypeFilter(value, condition = "==")
TxStartFilter(value, condition = "==")
TxEndFilter(value, condition = "==")
ProteinIdFilter(value, condition = "==")
UniprotFilter(value, condition = "==")
SeqNameFilter(value, condition = "==")
SeqStrandFilter(value, condition = "==")

\S4method{show}{AnnotationFilter}(object)

\S4method{condition}{AnnotationFilter}(object)

\S4method{value}{AnnotationFilter}(object)
}
\arguments{
\item{object}{An \code{AnnotationFilter} object}

\item{value}{Value for the filter.}

\item{condition}{character(1) defining the condition to be used in the filter.
For numeric/integer filters one of \code{"=="}, \code{"!="}, \code{">"},
\code{"<"}, \code{">="} and \code{"<="}. For character filter/values
\code{"=="}, \code{"!="}, \code{"startsWith"} and \code{"endsWith"} are
allowed. Default condition is "==".}
}
\description{
These functions are used to create filters for annotation
resources.

\code{supportedFilters()} lists all defined filters; by default filters are
only available for tables containing the field on which the filter
acts. See the vignette for a description to use filters for databases in which
the database table column name differs from the default `field` of the filter.

\code{condition} get the \code{condition} value for the filter
\code{object}.

\code{value} get or set the \code{value} for the filter
\code{object}.
}
\examples{
supportedFilters()

## Create a SymbolFilter to filter on a gene's symbol.
sf <- SymbolFilter("BCL2")
sf

## Create a GeneStartFilter to filter based on the genes' chromosomal start
## coordinates
gsf <- GeneStartFilter(10000, condition = ">")
gsf

}

