% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationFilter.R
\docType{class}
\name{supportedFilters}
\alias{AnnotationFilter-class}
\alias{CdsEndFilter}
\alias{CdsEndFilter-class}
\alias{CdsStartFilter}
\alias{CdsStartFilter-class}
\alias{EntrezFilter}
\alias{EntrezFilter-class}
\alias{ExonEndFilter}
\alias{ExonEndFilter-class}
\alias{ExonIdFilter}
\alias{ExonIdFilter-class}
\alias{ExonNameFilter}
\alias{ExonNameFilter-class}
\alias{ExonRankFilter}
\alias{ExonRankFilter-class}
\alias{ExonStartFilter}
\alias{ExonStartFilter-class}
\alias{GRangesFilter}
\alias{GRangesFilter-class}
\alias{GeneBiotypeFilter}
\alias{GeneBiotypeFilter-class}
\alias{GeneEndFilter}
\alias{GeneEndFilter-class}
\alias{GeneIdFilter}
\alias{GeneIdFilter-class}
\alias{GeneStartFilter}
\alias{GeneStartFilter-class}
\alias{GenenameFilter}
\alias{GenenameFilter-class}
\alias{ProteinIdFilter}
\alias{ProteinIdFilter-class}
\alias{SeqNameFilter}
\alias{SeqNameFilter-class}
\alias{SeqStrandFilter}
\alias{SeqStrandFilter-class}
\alias{SymbolFilter}
\alias{SymbolFilter-class}
\alias{TxBiotypeFilter}
\alias{TxBiotypeFilter-class}
\alias{TxEndFilter}
\alias{TxEndFilter-class}
\alias{TxIdFilter}
\alias{TxIdFilter-class}
\alias{TxNameFilter}
\alias{TxNameFilter-class}
\alias{TxStartFilter}
\alias{TxStartFilter-class}
\alias{UniprotFilter}
\alias{UniprotFilter-class}
\alias{condition}
\alias{condition,AnnotationFilter-method}
\alias{condition,GRangesFilter-method}
\alias{show,AnnotationFilter-method}
\alias{show,GRangesFilter-method}
\alias{supportedFilters}
\alias{value}
\alias{value,AnnotationFilter-method}
\alias{value,GRangesFilter-method}
\title{Filters for annotation objects}
\usage{
CdsStartFilter(value, condition = "==")
CdsEndFilter(value, condition = "==")
ExonIdFilter(value, condition = "==")
ExonNameFilter(value, condition = "==")
ExonRankFilter(value, condition = "==")
ExonStartFilter(value, condition = "==")
ExonEndFilter(value, condition = "==")
GeneIdFilter(value, condition = "==")
GenenameFilter(value, condition = "==")
GeneBiotypeFilter(value, condition = "==")
GeneStartFilter(value, condition = "==")
GeneEndFilter(value, condition = "==")
EntrezFilter(value, condition = "==")
SymbolFilter(value, condition = "==")
TxIdFilter(value, condition = "==")
TxNameFilter(value, condition = "==")
TxBiotypeFilter(value, condition = "==")
TxStartFilter(value, condition = "==")
TxEndFilter(value, condition = "==")
ProteinIdFilter(value, condition = "==")
UniprotFilter(value, condition = "==")
SeqNameFilter(value, condition = "==")
SeqStrandFilter(value, condition = "==")

GRangesFilter(value, condition = "overlapping", feature = "gene")

\S4method{show}{AnnotationFilter}(object)

\S4method{condition}{AnnotationFilter}(object)

\S4method{value}{AnnotationFilter}(object)

\S4method{show}{GRangesFilter}(object)

\S4method{condition}{GRangesFilter}(object)

\S4method{value}{GRangesFilter}(object)
}
\arguments{
\item{value}{Value for the filter. For \code{GRangesFilter} a
\code{\link[GenomicRanges]{GRanges}} object.}

\item{condition}{character(1) defining the condition to be used in the filter.
For numeric/integer filters one of \code{"=="}, \code{"!="}, \code{">"},
\code{"<"}, \code{">="} and \code{"<="}. For character filter/values
\code{"=="}, \code{"!="}, \code{"startsWith"} and \code{"endsWith"} are
allowed. Default condition is "==". For \code{GRangesFilter} it can be
\code{"within"} (for the feature to be completely within the range) or
\code{"overlapping"}, for the feature to be (partially) overlapping with the
range.}

\item{feature}{\code{character(1)} defining on what feature the
\code{GRangesFilter} should be applied. Choices could be \code{"gene"},
\code{"tx"} or \code{"exon"}.}

\item{object}{An \code{AnnotationFilter} or \code{GRangesFilter} object}
}
\description{
The filters extending the base \code{AnnotationFilter} class
represent a simple filtering concept for annotation resources.
Each filter object is thought to filter on a single (database) table column
based on the provided values and the defined condition.

Filter instances should be created using the constructor functions (e.g.
\code{GeneIdFilter}) and not by calls to \code{new}.

\code{supportedFilters()} lists all defined filters.

\code{condition} get the \code{condition} value for the filter
\code{object}.

\code{value} get or set the \code{value} for the filter
\code{object}.
}
\details{
By default filters are only available for tables containing the
field on which the filter acts (i.e. that contain a column with the name
matching the value of the \code{field} slot of the object). See the vignette
for a description to use filters for databases in which the database table
column name differs from the default \code{field} of the filter.
}
\examples{
supportedFilters()

## Create a SymbolFilter to filter on a gene's symbol.
sf <- SymbolFilter("BCL2")
sf

## Create a GeneStartFilter to filter based on the genes' chromosomal start
## coordinates
gsf <- GeneStartFilter(10000, condition = ">")
gsf

## filter by GRanges
GRangesFilter(as("chr10:87869000-87876000", "GRanges"))

}

