% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationFilterList.R
\docType{class}
\name{AnnotationFilterList}
\alias{.AnnotationFilterList}
\alias{AnnotationFilterList}
\alias{AnnotationFilterList-class}
\alias{logOp}
\alias{logOp,AnnotationFilterList-method}
\alias{show,AnnotationFilterList-method}
\alias{value,AnnotationFilterList-method}
\title{Combining AnnotationFilters}
\usage{
AnnotationFilterList(..., logOp = character())

AnnotationFilterList(..., logOp = character())

\S4method{show}{AnnotationFilterList}(object)

\S4method{logOp}{AnnotationFilterList}(object)

\S4method{value}{AnnotationFilterList}(object)
}
\arguments{
\item{...}{individual \code{\link{AnnotationFilter}} objects or a mixture of
\code{AnnotationFilter} and \code{AnnotationFilterList} objects.}

\item{logOp}{\code{character} of length being equal to the numner of submitted
\code{AnnotationFilter} objects -1. Each value representing the logical
operation to combine consecutive filters, i.e. the first element being the
logical operation to combine the first and second \code{AnnotationFilter}, the
second element being the logical operation to combine the second and third
\code{AnnotationFilter} and so on. Allowed values are \code{"&"} and
\code{"|"}. The function assumes a logical \emph{and} between all elements by
default.}

\item{object}{An \code{AnnotationFilterList}.}
}
\description{
The \code{AnnotationFilterList} allows to combine filter objects
extending the \code{\link{AnnotationFilter}} class to construct more complex
qyeries. Consecutive filter objects in the \code{AnnotationFilterList} can be
combined by a logical \emph{and} (\code{&}) or \emph{or} (\code{|}).

\code{logOp()} get the \code{character} with the logical
operators combining the filters.

\code{value()} get a \code{list} with the \code{AnnotationFilter}
objects.
}
\examples{

## Create some AnnotationFilters
gf <- GenenameFilter(c("BCL2", "BCL2L11"))
tbtf <- TxBiotypeFilter("protein_coding", condition = "!=")

## Combine both to an AnnotationFilterList. By default elements are combined
## using a logical "and" operator. The filter list represents thus a query
## like: get all features where the gene name is either ("BCL2" or "BCL2L11")
## and the transcript biotype is not "protein_coding".
afl <- AnnotationFilterList(gf, tbtf)
afl

## Get the logical operator combining the filters
logOp(afl)

## and the list with the filter objects
value(afl)

## Create a filter in the form of: get all features where the gene name is
## either ("BCL2" or "BCL2L11") and the transcript biotype is not
## "protein_coding" or the seq_name is "Y". Hence, this will get all feature
## also found by the previous AnnotationFilterList and returns also all
## features on chromosome Y.
afl <- AnnotationFilterList(gf, tbtf, SeqNameFilter("Y"), logOp = c("&", "|"))
afl

}
\seealso{
\code{\link{supportedFilters}} for available
\code{\link{AnnotationFilter}} objects
}

